Move contract_to_tensor and ubersum into pyro.ops
